# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PetWorld API",
    "description": "API 測試頁面",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://petworld.herokuapp.com/",
      "description": "Heroku"
    }
  ],
  "tags": [
    {
      "name": "ANIMAL",
      "description": "Everything about animal",
      "externalDocs": {
        "description": "Find out more",
        "url": "https://data.gov.tw/"
      }
    },
    {
      "name": "DONATE",
      "description": "Access to Petstore orders"
    },
    {
      "name": "user",
      "description": "Operations about user",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/auth/testAPI": {
      "get": {
        "summary": "API連線測試",
        "description": "確認連線是否正常",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },    
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "登入系統",
        "operationId": "loginUser",
        "requestBody":{
          "content":{
            "application/json":{
              "schema": {
                "type": "object",
                "properties":{
                  "amount":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  }
                }
              }
            }
          }
      },"responses": {
        "200": {
          "description": "成功！",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "400": {
          "description": "Error username/password",
          "content": {}
        }
      }
      }
    },
    "/user/create": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "用戶註冊",
        "operationId": "signupUser",
        "requestBody":{
          "content":{
            "application/json":{
              "schema": {
                "type": "object",
                "properties":{
                  "account":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  },
                  "email":{
                    "type":"string"
                  },
                  "phone":{
                    "type":"number"
                  },
                  "address":{
                    "type":"string"
                  },
                  "job":{
                    "type":"string"
                  },
                  "age":{
                    "type":"number"
                  }
                }
              }

            }
          }

      },"responses": {
        "200": {
          "description": "成功註冊！",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "400": {
          "description": "Error",
          "content": {}
        }
      }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "ADMIN"
        ],
        "summary": "列出所有已註冊用戶",
        "operationId": "findAllUser",
        "responses": {
          "200": {
            "description": "成功！"
          },
          "400": {
            "description": "連線問題",
            "content": {}
          }
        }
      }
    },
    "/user/{email}": {
      "get": {
        "tags": [
          "ADMIN"
        ],
        "summary": "使用{email}找出特定用戶",
        "operationId": "findSpecificUser",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "輸入使用者電子郵件",
            "required": true,
            "schema": {
              "type": "string",
              "properties": {
                "name": {
                  "type": "string"
                  }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功！"
          },
          "400": {
            "description": "連線問題",
            "content": {}
          }
        }
      }
    },
    "/users/viewadopt": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "列出領養紀錄",
        "operationId": "useradopt",
        "requestBody":{
          "content":{
            "application/json":{
              "schema": {
                "type": "object",
                "properties":{
                  "userId":{
                    "type":"number"
                  }
                }
              }
            }
          }
      },
        "responses": {
          "200": {
            "description": "成功！"
          },
          "400": {
            "description": "連線問題",
            "content": {}
          }
        }
      }
    },
    "/users/viewdonate": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "列出捐贈紀錄",
        "operationId": "userdonate",
        "requestBody":{
          "content":{
            "application/json":{
              "schema": {
                "type": "object",
                "properties":{
                  "userId":{
                    "type":"number"
                  }
                }
              }
            }
          }
      },
        "responses": {
          "200": {
            "description": "成功！"
          },
          "400": {
            "description": "連線問題",
            "content": {}
          }
        }
      }
    },
    "/users/logout": {
      "get": {
        "tags": [
          "USER"
        ],
        "summary": "登出",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "登出"
          },
          "400": {
            "description": "發生問題",
            "content": {}
          }
        }
      }
    },
    "/donate": {
        "post": {
          "tags": [
            "DONATE"
          ],
          "summary": "捐贈",
          "operationId": "donate",
          "requestBody":{
            "content":{
              "application/json":{
                "schema": {
                  "type": "object",
                  "properties":{
                    "userId":{
                      "type":"number"
                    },
                    "dog":{
                      "type":"number"
                    },
                    "putty":{
                      "type":"number"
                    },
                    "cat":{
                      "type":"number"
                    },
                    "kitty":{
                      "type":"number"
                    },
                    "shelter":{
                      "type":"string"
                    }
                  }
                }
              }
            }
        },"responses": {
          "200": {
            "description": "捐贈成功！",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {}
          }
        }
        }
    },
    "/animal": {
      "get": {
        "tags": [
          "ANIMAL"
        ],
        "summary": "列出收容所動物",
        "operationId": "findAnimal",
        "responses": {
          "200": {
            "description": "成功！"
          },
          "400": {
            "description": "連線問題",
            "content": {}
          }
        }
      }
    },
    "/animal/:animalId/adopt": {
      "post": {
        "tags": [
          "ANIMAL"
        ],
        "summary": "領養動物",
        "operationId": "adopt",
        "parameters": [
          {
            "in": "path",
            "name": "animalId",
            "description": "輸入animalId",
            "required": true,
            "schema": {
              "type": "string",
              "properties": {
                "name": {
                  "type": "string"
                  }
              }
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema": {
                "type": "object",
                "properties":{
                  "userId":{
                    "type":"number"
                  }
                }
              }

            }
          }

      },"responses": {
        "200": {
          "description": "領養成功！",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "400": {
          "description": "領養失敗",
          "content": {}
        }
      }
      }
    },
    "/animal/{animalId}": {
      "get": {
        "tags": [
          "ANIMAL"
        ],
        "summary": "顯示單一動物資料",
        "operationId": "findanimal",
        "parameters": [
          {
            "in": "path",
            "name": "animalId",
            "description": "輸入animalId",
            "required": true,
            "schema": {
              "type": "string",
              "properties": {
                "name": {
                  "type": "string"
                  }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "動物資料已列出！"
          },
          "400": {
            "description": "查無資料",
            "content": {}
          }
        }
      }
    }
  }
}